arcgis api升级 添加时间轴 bug修复

+++ src/main/java/com/clb/entity/InspectionRecord.java
@@ -17,7 +17,7 @@ public class InspectionRecord {
     private int id;
     private Integer glxmId;
     private String xcryName;
-    private float xcdate;
+    private long xcdate;
     private String xcjl;
     private String xchdUid;
     private Integer addUid;
